use global directory for root view
